Clean project and reorganize directories.
Remove unused code from other course exercises.
Make recipe projet more readable.
