+++ README.md
@@ -1,4 +1,4 @@
-# PeopleDogsApp
+# Angular courses: A recipe app.
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 7.3.1.
