+++ src/app/app-routing.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import {MealsComponent} from "./meals/meals.component";
-import {MealListComponent} from "./meals/meal-list/meal-list.component";
-import {MealRecipeComponent} from "./meals/meal-recipe/meal-recipe.component";
+import {MealListComponent} from "./meal-list/meal-list.component";
+import {MealRecipeComponent} from "./meal-recipe/meal-recipe.component";
 
 const appRoutes: Routes = [
   {path: '', component: MealsComponent},
   {path: 'meals', component: MealsComponent, children: [
-    {path: ':type', component: MealListComponent, children: [
+    {path: ':category', component: MealListComponent, children: [
         {path: ':name', component: MealRecipeComponent}
       ]}
   ]},
