+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from '@angular/core';
+import {Component} from '@angular/core';
 
 @Component({
   selector: 'app-root',
@@ -6,26 +6,4 @@ import {Component, OnInit} from '@angular/core';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-  username: string = 'you';
-  clicks = [];
-  hasClick: boolean = false;
-
-  numbers: number[] = [];
-
-  ngOnInit() {
-  }
-
-  resetUsername() {
-    this.username = '';
-  }
-
-  onClick() {
-    this.hasClick = true;
-    this.clicks.push(this.clicks.length + 1);
-  }
-
-  onStartInterval(myIntervalNumber: number) {
-    console.log(myIntervalNumber);
-    this.numbers.push(myIntervalNumber);
-  }
 }
