--- src/app/even/even.component.html
@@ -1,3 +1,0 @@
-<p>
-  even works!
-</p>
