--- src/app/even/even.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-even',
-  templateUrl: './even.component.html',
-  styleUrls: ['./even.component.scss']
-})
-export class EvenComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
