--- src/app/game-control/game-control.component.html
@@ -1,3 +1,0 @@
-<button class="btn btn-primary" (click)="startGame()">Start</button>
-&nbsp;
-<button class="btn btn-primary" (click)="stopGame()">Stop</button>
