+++ src/app/header/header.component.html
@@ -18,9 +18,24 @@
           Meals
         </a>
         <ul class="dropdown-menu">
-          <li routerLinkActive="active"><a class="dropdown-item" [routerLink]="['/meals', 'mexican']">Mexican</a></li>
-          <li routerLinkActive="active"><a class="dropdown-item" [routerLink]="['/meals', 'italian']">Italian</a></li>
-          <li routerLinkActive="active"><a class="dropdown-item" [routerLink]="['/meals', 'greek']">Greek</a></li>
+          <li routerLinkActive="active">
+            <a class="dropdown-item"
+               [routerLink]="['/meals', 'mexican']">
+              Mexican
+            </a>
+          </li>
+          <li routerLinkActive="active">
+            <a class="dropdown-item"
+               [routerLink]="['/meals', 'italian']">
+              Italian
+            </a>
+          </li>
+          <li routerLinkActive="active">
+            <a class="dropdown-item"
+               [routerLink]="['/meals', 'greek']">
+              Greek
+            </a>
+          </li>
         </ul>
       </li>
     </ul>
