+++ src/app/header/header.component.ts
@@ -6,5 +6,4 @@ import {Component} from "@angular/core";
   styleUrls: ['./header.component.scss'],
 })
 export class headerComponent {
-  
 }
