+++ src/app/meal-item/meal-item.component.html
@@ -6,9 +6,6 @@
   </div>
 
   <div class="meal__infos">
-    <div class="meal__info__allergen">
-      <app-meal-allergene></app-meal-allergene>
-    </div>
     <h3 class="meal__info__title">
       {{ mealItem.name }}
     </h3>
