+++ src/app/meal-item/meal-item.component.scss
