+++ src/app/meal-list/meal-list.component.html
@@ -1,0 +1,9 @@
+<h2 class="meal-list-title">
+  {{mealCategory}}
+</h2>
+
+<div class="meal-list row">
+  <app-meal-item class="meal-list__item col-xs-12 col-sm-4" *ngFor="let meal of meals[mealCategory]" [mealItem]="meal"></app-meal-item>
+</div>
+
+<router-outlet></router-outlet>
