--- src/app/meals/meal-list/meal-item/meal-allergene/meal-allergene.component.html
@@ -1,1 +1,0 @@
-Beef + Gluten free
