--- src/app/meals/meal-list/meal-item/meal-allergene/meal-allergene.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-meal-allergene',
-  templateUrl: './meal-allergene.component.html',
-  styleUrls: ['./meal-allergene.component.scss'],
-})
-export class MealAllergeneComponent {
-
-}
