--- src/app/meals/meal-list/meal-list.component.html
@@ -1,8 +1,0 @@
-<h2 class="meal-list-title">
-  {{mealType}}
-</h2>
-
-<div class="meal-list row">
-  <app-meal-item class="meal-list__item col-xs-12 col-sm-4" *ngFor="let meal of meals[mealType]" [mealItem]="meal"></app-meal-item>
-  <router-outlet></router-outlet>
-</div>
