+++ src/app/models/meal.model.ts
