--- src/app/odd/odd.component.html
@@ -1,3 +1,0 @@
-<p>
-  odd - {{ number }}
-</p>
